Type adjustments

+++ nunchuckUltimatum/nunchuckUltimatum/boot.cpp
@@ -7,9 +7,14 @@ void drawboot(MI0283QT9 lcd, int x, int y) {
 	uint8_t lengte = 25;
 	uint8_t stapgroote = 3;
 	uint8_t stap = 1;
+	//draw square
 	lcd.drawRect(x, y, breete, lengte, RGB(0, 0, 0));
+	//draws line around the square
 	lcd.fillRect(x + 1, y + 1, breete - 2, lengte - 2, RGB(156, 102, 31));
-	int mastPositie = -2;
+	//can place the mast further backward on the boat
+	int8_t mastPositie = -2;
+	// draws mast
 	lcd.fillCircle(x + breete / 2, y + lengte / 2 + mastPositie, (breete + lengte) / 12, RGB(82, 82, 82));
+	//draws sail
 	lcd.fillTriangle(x, y + lengte / 2 - (breete + lengte) / 12 + mastPositie, x + breete - 1, y + lengte / 2 - (breete + lengte) / 12 + mastPositie, x + breete / 2, y - 2 + lengte / 2 - (breete + lengte) / 12 + mastPositie, RGB(255, 255, 255));
 }
